+++ packages/block-library/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 -   Replaced `classnames` package with the faster and smaller `clsx` package ([#61138](https://github.com/WordPress/gutenberg/pull/61138)).
 
+## 8.34.0 (2024-05-02)
+
 ## 8.33.0 (2024-04-19)
 
 ## 8.32.0 (2024-04-03)